Fix: in-memory-database lost within container-image-scanner
Based on the limitation of the SqlitePool it can not guarantue to hold
at least one connection although it is configured that way.

That's why container-image-scanner also needs to utilize
SqliteConnectionContainer, like in openvasd scheduler, artificially
although it does not need it from a functional point of view.
